user info validation

+++ server/api/routes/user.js
@@ -1,10 +1,14 @@
 const express = require("express");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
+const { check, validationResult } = require('express-validator');
 
 const authorize = require("../middleware/check-auth");
 const isAdmin = require("../middleware/check-isAdmin");
 const router = express.Router();
+
+const validateUserLogin = require("../middleware/validateUserLogin")
+const validateInput = require("../middleware/validateCreateUser");
 const model = require('../../models/index');
 
 const hashCount = 12;
@@ -13,8 +17,13 @@ router.get('/users',[authorize, isAdmin], (req, res) => {
   model.User.findAll().then(users => res.json(users));
 });
 
-router.post("/signup", (req, res, next) => {
-  
+router.post("/signup", [validateInput], (req, res, next) => {
+
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
   model.User.findAll({ 
     
     where: { username: req.body.username } 
@@ -55,8 +64,13 @@ router.post("/signup", (req, res, next) => {
 
 });
 
-router.post("/login", (req, res, next) => {
+router.post("/login", [validateUserLogin], (req, res, next) => {
   
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
   model.User.findOne({ 
     
     where: { username: req.body.username } 
@@ -135,7 +149,19 @@ router.post("/login", (req, res, next) => {
 
 });
 
-router.delete("/delete/:userId",  [authorize, isAdmin], (req, res, next) => {
+router.delete("/delete/:userId",  [
+  check("userId").not().isEmpty().withMessage("userId not provided")
+  .trim()
+  .escape()
+  .isNumeric().withMessage("not a number"),
+  authorize, 
+  isAdmin
+], (req, res, next) => {
+
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
 
   model.User.destroy({
